# Added scroller to trait page and changed sample tables to use it