--- a/wqflask/wqflask/show_trait/show_trait.py
+++ b/wqflask/wqflask/show_trait/show_trait.py
@@ -275,7 +275,9 @@ class ShowTrait(object):
                        data_scale = self.dataset.data_scale,
                        sample_group_types = self.sample_group_types,
                        sample_lists = sample_lists,
-                       attribute_names = self.sample_groups[0].attributes,
+                       se_exists = self.sample_groups[0].se_exists,
+                       has_num_cases = self.has_num_cases,
+                       attributes = self.sample_groups[0].attributes,
                        categorical_vars = ",".join(categorical_var_list),
                        num_values = self.num_values,
                        qnorm_values = self.qnorm_vals,
@@ -454,6 +456,7 @@ class ShowTrait(object):
         self.primary_sample_names = primary_sample_names
         self.dataset.group.allsamples = all_samples_ordered
 
+
 def quantile_normalize_vals(sample_groups):
     def normf(trait_vals):
         ranked_vals = ss.rankdata(trait_vals)
@@ -492,6 +495,7 @@ def quantile_normalize_vals(sample_groups):
 
     return qnorm_by_group
 
+
 def get_z_scores(sample_groups):
     zscore_by_group = []
     for sample_type in sample_groups:
@@ -516,6 +520,7 @@ def get_z_scores(sample_groups):
 
     return zscore_by_group
 
+
 def get_nearest_marker(this_trait, this_db):
     this_chr = this_trait.locus_chr
     logger.debug("this_chr:", this_chr)
@@ -539,6 +544,7 @@ def get_nearest_marker(this_trait, this_db):
     else:
         return result[0][0]
 
+
 def get_table_widths(sample_groups, has_num_cases=False):
     stats_table_width = 250
     if len(sample_groups) > 1:
@@ -555,6 +561,7 @@ def get_table_widths(sample_groups, has_num_cases=False):
 
     return stats_table_width, trait_table_width
 
+
 def has_num_cases(this_trait):
     has_n = False
     if this_trait.dataset.type != "ProbeSet" and this_trait.dataset.type != "Geno":
@@ -564,6 +571,7 @@ def has_num_cases(this_trait):
                 break
 
     return has_n
+
 
 def get_trait_units(this_trait):
     unit_type = ""
@@ -584,6 +592,7 @@ def get_trait_units(this_trait):
 
     return unit_type
 
+
 def check_if_attr_exists(the_trait, id_type):
     if hasattr(the_trait, id_type):
         if getattr(the_trait, id_type) == None or getattr(the_trait, id_type) == "":
@@ -592,6 +601,7 @@ def check_if_attr_exists(the_trait, id_type):
             return True
     else:
         return False
+
 
 def get_ncbi_summary(this_trait):
     if check_if_attr_exists(this_trait, 'geneid'):
@@ -604,6 +614,7 @@ def get_ncbi_summary(this_trait):
             return None
     else:
         return None
+
 
 def get_categorical_variables(this_trait, sample_list):
     categorical_var_list = []
@@ -623,6 +634,7 @@ def get_categorical_variables(this_trait, sample_list):
 
     return categorical_var_list
 
+
 def get_genotype_scales(genofiles):
     geno_scales = {}
     if type(genofiles) is list:
@@ -633,6 +645,7 @@ def get_genotype_scales(genofiles):
         geno_scales[genofiles] = get_scales_from_genofile(genofiles)
 
     return geno_scales
+
 
 def get_scales_from_genofile(file_location):
     geno_path = locate_ignore_error(file_location, 'genotype')
@@ -685,6 +698,7 @@ def get_scales_from_genofile(file_location):
             else:
                 if i > first_marker_line + 10:
                     break
+
 
     #ZS: This assumes that both won't be all zero, since if that's the case mapping shouldn't be an option to begin with
     if mb_all_zero: