--- a/wqflask/wqflask/static/new/css/show_trait.css
+++ b/wqflask/wqflask/static/new/css/show_trait.css
@@ -47,3 +47,11 @@ table.dataTable.cell-border tbody tr th:first-child,
 table.dataTable.cell-border tbody tr td:first-child {
   border-left: 1px solid #ccc;
 }
+
+.checkbox {
+  min-height: 20px;
+}
+
+.trait_value_input {
+  text-align: right;
+}