… wqflask/wqflask/static/new/javascript/show_trait.js
… wqflask/wqflask/static/new/javascript/show_trait_mapping_tools.js
… wqflask/wqflask/templates/show_trait.html
… wqflask/wqflask/templates/show_trait_transform_and_filter.html