@BonfaceKilz 2020-09-16
zsloan <notifications@github.com> writes:

> Ah, I wasn't planning on keeping that there, but I guess there's no reason
> to not delete it since the old code is still accessible.

I normally try to get into the habit of cleaning
things up as soon as I can, because I forget :)

> On Tue, Sep 15, 2020 at 5:52 PM BonfaceKilz <notifications@github.com>
> wrote:
>
>> *@BonfaceKilz* commented on this pull request.
>> ------------------------------
>>
>> In wqflask/wqflask/templates/show_trait.html
>> <https://github.com/genenetwork/genenetwork2/pull/442#discussion_r489048143>
>> :
>>
>> > + // 'processing': true,
>> + // 'language': {
>> + // 'loadingRecords': '&nbsp;',
>> + // 'processing': 'Loading...'
>> + // }
>>
>> *suggestion: nitpick:* It's a good idea to remove unnecessary comments.
>> This class of comments--commented out code-- after a while becomes "dead
>> code" that either usually ends either being clutter or outright misleading.
>> ------------------------------
>>
>> In wqflask/wqflask/templates/show_trait.html
>> <https://github.com/genenetwork/genenetwork2/pull/442#discussion_r489048286>
>> :
>>
>> > + "scrollY": "50vh",
>> + 'scroller': true,
>> + 'scrollCollapse': true
>> + // 'processing': true,
>> + // 'language': {
>> + // 'loadingRecords': '&nbsp;',
>> + // 'processing': 'Loading...'
>> + // }
>>
>> See above.
>>
>> —
>> You are receiving this because you authored the thread.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/genenetwork/genenetwork2/pull/442#pullrequestreview-489160631>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/AANQJGDQGSRABIWCS7IDNI3SF7V3TANCNFSM4RNPZ56A>
>> .
>>
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or unsubscribe.*

-- 
Bonface M. K. (https://www.bonfacemunyoki.com)
Chief Emacs Mchochezi
GPG key = D4F09EB110177E03C28E2FE1F5BBAE1E0392253F