@zsloan 2020-09-16
After spending some time with this, one of the main issues seems to be making the DataTables instances accessible to the show_trait.js and show_trait_mapping_tools.js files. I think the easiest way to do this might be to put the event listeners in the template JS and have those pass the DataTables instances to the functions that do whatever is supposed to happen (updating figures, hiding rows, etc). Otherwise I think a bunch of existing template logic would need to be converted to JS and various template variables would need to be made accessible to the JS code (which I think has a cost since it's basically adding to what has to be loaded by the browser).