@BonfaceKilz 2020-09-16
zsloan <notifications@github.com> writes:

> After spending some time with this, one of the main issues seems to be making the DataTables instances
> accessible to the show_trait.js and show_trait_mapping_tools.js files. I think the easiest way to do this
> might be to put the event listeners in the template JS and have those pass the DataTables instances to the
> functions that do whatever is supposed to happen (updating figures, hiding rows, etc). Otherwise I think a
> bunch of existing template logic would need to be converted to JS and various template variables would need to
> be made accessible to the JS code (which I think has a cost since it's basically adding to what has to be
> loaded by the browser).
>

Yeah. Go for something that works. Then worry
about the costs later.

> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or unsubscribe.*

-- 
Bonface M. K. (https://www.bonfacemunyoki.com)
Chief Emacs Mchochezi
GPG key = D4F09EB110177E03C28E2FE1F5BBAE1E0392253F