@zsloan 2020-09-23
Just as an update on this, I'm running into some problems related to the speed with which the DataTables API pulls values out of the cells of a large table. It's not a big deal if a table just has <500 rows, but for something like the HSNIH-Palmer data it's prohibitively slow any time a table value is changed (or a transform is run, etc).

I'm not sure how to address this (or if it even can be reasonably addressed without some very significant change to the way this works). I think that all of this data would need to be stored outside of the table and updated when specific rows/cells are changed, because I think it that it's probably always going to be very slow to use the DataTables API to pull from all table rows. It would probably be faster if it was just pulling the table data, but because of the text inputs it has to actually pull the rendered cell contents.