@zsloan 2020-09-24
I think I managed to solve the issue of speed. It actually seems significantly faster than it was before using Scroller now (for pulling data out of the table).