@BonfaceKilz 2020-09-25
zsloan <notifications@github.com> writes:

> Just as an update on this, I'm running into some problems related to the speed with which the DataTables API
> pulls values out of the cells of a large table. It's not a big deal if a table just has <500 rows, but for
> something like the HSNIH-Palmer data it's prohibitively slow any time a table value is changed (or a transform
> is run, etc).
>
> I'm not sure how to address this (or if it even can be reasonably addressed without some very significant
> change to the way this works). I think that all of this data would need to be stored outside of the table and
> updated when specific rows/cells are changed, because I think it that it's probably always going to be very
> slow to use the DataTables API to pull from all table rows. It would probably be faster if it was just pulling
> the table data, but because of the text inputs it has to actually pull the rendered cell contents.
>

Perhaps using some web-storage API is of
consideration? Though IMO that would make things
really complicated. I'd leave this open, and first
prioritize merging Udit's work. We can figure
something out in due time.

> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or unsubscribe.*

-- 
Bonface M. K. (https://www.bonfacemunyoki.com)
Chief Emacs Mchochezi / Twitter: @BonfaceKilz
GPG key = D4F09EB110177E03C28E2FE1F5BBAE1E0392253F