@zsloan 2020-09-25
I was initially mistaking the speed with Scroller + my initial use of DataTables API for pre-Scroller, though it's still a bit faster than pre-Scroller
The page being tested on - http://gn2-zach.genenetwork.org/show_trait?trait_id=10443&dataset=HSNIH-PalmerPublish

With Scroller and first DataTables API use (that iterated through column cells) - 45.09458999999333 seconds
With Scroller and current DataTables API use (that just looks at individual columns and only does the "to JQuery" conversion on individual cells) - 0.12849000000278465 seconds
Before Scroller and switch to DataTables API - 0.17568499999470077 seconds

More noticeable is the amount of time it takes to block a range of samples by index (in this case blocking samples 1000-1100):

Before Scroller - 1.8479800000059186 seconds (also the page in general is sluggish here, with typing into text fields having lag)
With Scroller - 0.12387499999022111 seconds